PINK: Rename director.h/cpp to surface.h/cpp
