+++ engines/pink/surface.cpp
@@ -26,7 +26,7 @@
 
 #include "pink/pink.h"
 #include "pink/cel_decoder.h"
-#include "pink/director.h"
+#include "pink/surface.h"
 #include "pink/objects/actions/action_sound.h"
 #include "pink/objects/actions/action_cel.h"
 #include "pink/objects/actions/action_text.h"
